Extract logic for pluggable search bar controls.
